requirements.txt
google-genai
pillow
numpy
torch